+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add generic templating
 - Add Flow\Driver\RevoltDriver
+- Add Flow\Driver\SpatieDriver
 - Add more quality tools from https://github.com/IngeniozIT/php-skeleton
 
 ## v1.1.3